+++ src/dlite-datamodel.h
@@ -26,7 +26,7 @@
 typedef struct _DLiteDataModel DLiteDataModel;
 
 
-#include "dlite-types.h"
+#include "dlite-type.h"
 #include "dlite-storage.h"
 #include "dlite-plugins.h"
 
@@ -88,7 +88,7 @@ int dlite_datamodel_get_dimension_size(const DLiteDataModel *d, const char *name
  */
 int dlite_datamodel_get_property(const DLiteDataModel *d, const char *name,
                                  void *ptr, DLiteType type, size_t size,
-                                 int ndims, const int *dims);
+                                 size_t ndims, const size_t *dims);
 
 /** @} */
 
@@ -116,7 +116,7 @@ int dlite_datamodel_get_property(const DLiteDataModel *d, const char *name,
 */
 int dlite_datamodel_set_property(DLiteDataModel *d, const char *name,
                                  const void *ptr, DLiteType type, size_t size,
-                                 int ndims, const int *dims);
+                                 size_t ndims, const size_t *dims);
 
 
 /**
